优化购物车删除单个商品代码

+++ sky-server/src/main/java/com/sky/service/impl/ShoppingCartServiceImpl.java
@@ -101,7 +101,15 @@ public class ShoppingCartServiceImpl implements ShoppingCartService {
                 .eq(shoppingCartDTO.getDishFlavor() != null, ShoppingCart::getDishFlavor, shoppingCartDTO.getDishFlavor())
                 .eq(ShoppingCart::getUserId, BaseContext.getCurrentId())
                 .eq(shoppingCartDTO.getSetmealId() != null, ShoppingCart::getSetmealId, shoppingCartDTO.getSetmealId());
-        shoppingCartDao.delete(lqw);
+
+        List<ShoppingCart> list = shoppingCartDao.selectList(lqw);
+        ShoppingCart cart = list.get(0);
+        if(cart.getNumber()>1){
+            cart.setNumber(cart.getNumber()-1);
+            shoppingCartDao.updateById(cart);
+        }else {
+            shoppingCartDao.delete(lqw);
+        }
     }
 
     /**
